refactor in order to expose mutateOffline
removed the ApolloStoreOffline file, now the store is created in the client's constructor

created the public function "mutateOffline" in the client and used by the client when the "mutate" is performed and the application is offline

type refactors, they are now all managed in the ApolloOfflineTypes file and exported to the index file
