+++ packages/apollo-offline/src/ApolloClientIDB.ts
@@ -1,11 +1,9 @@
 import { ICacheStorage, CacheOptions } from '@wora/cache-persist';
 import IDBStorage, { IOnUpgrade } from '@wora/cache-persist/lib/idbstorage';
-import ApolloClientOffline, { OfflineApolloClientOptions } from './ApolloClientOffline';
+import ApolloClientOffline from './ApolloClientOffline';
 import { InMemoryCacheConfig } from 'apollo-cache-inmemory';
 import ApolloCache from '@wora/apollo-cache';
-
-type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
-type ApolloClientIDBOptions = Omit<OfflineApolloClientOptions, 'cache'>;
+import { ApolloClientIDBOptions } from './ApolloOfflineTypes';
 
 class ApolloClientIDB {
     public static create(
